Updated Naming convesion to match student frontend
