+++ frontend/src/pages/home/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Container, Button } from "@material-ui/core";
 import { Link } from "react-router-dom";
 import { AuthContext } from "../../context/auth";
-import TeacherSubjectListPage from "../teacher/subjectList";
+import TeacherSubjectListPage from "../teacherSubjectList";
 
 export default function HomePage() {
   const { authState, setAuthState } = React.useContext(AuthContext);
